refactor ActivityDetailScreen
